Move `check_opencast_login` to new `auth::opencast` module
